requirements.txt
langchain-community==0.4.1
langchain-openai==1.1.6
langchain-core==1.2.6
langchain-text-splitters==1.1.0
langchain-chroma==1.1.0
chromadb==1.4.0
openai==2.14.0
python-dotenv==1.0.1
unstructured==0.14.4
nltk==3.9.2
tiktoken==0.7.0
pydantic==2.12.5
pydantic-settings==2.12.0
httpx==0.28.1
httpx-sse==0.4.3
sentence-transformers==2.7.0
google-generativeai==0.8.3